переходим на ООП

+++ doodle/doodle/Map.h
@@ -2,5 +2,5 @@
 
 using namespace sf;
 
-const int HEIGHT_MAP = 25;
-const int WIDTH_MAP = 40;
+const float HEIGHT_MAP = 25;
+const float WIDTH_MAP = 40;
